Fix since error on short lived containers
If the container is short lived it will terminate before cl inspects the
container thus using the finished time even when the start time is more
appropriate. By using the earliest of the two all appropriate logs are
included. (The finished time is used when containers are restarted;
including everything from last termination)