Remove retry mechanism
It's unclear if this is needed. It's also unclear which errors should be
retried. ErrConflict is returned when tailing a container that's already
been removed, something which may happen rather frequently for
auto-removed short lived containers, and should probably not be retried.
Unauthorized and forbidden is no use retried either. It can be readded
and reevaluated if needed.