feat: Keep track of seen containers to set "Since" time more accurately
Before, logs where always truncated to latest container start. To get rid
of this restriction containers are kept track of, not just if their
active, but ever (as long as they fit in cache and are not removed) Log
streams end when containers terminate and if the container starts again
using the --since time will print logs that were already printed on
earlier starts of the container. If the container has been seen earlier
logs will be printed from latest start/finish time. Unless --since is
later than the start time which is unlikely unless retrying a failed log
stream is done (which is not merged yet).