feat: Use resumeRequest on container restart
The two previous methods of determining where to restart a log stream do
not work well enough. Using the container start time sometimes skip
the first logs, using the latest stop time sometimes include logs from
previous run that were already printed.

So this new method keeps track of which logs are already printed and
skips them. It stores the printed log's timestamp and the number of logs
printed on that timestamp.

This introduces a new order dependency though. If a container delete
event is handled before the log stream of said container terminates it
will add an entry to the `seenContainers` cache even though the
container does not exist. Haven't seen that happen though and the only
downside is it might evict containers from that cache that still exist.