fix: Align template variables with Docker nomenclature
The .ContainerName was in fact the service name when using compose. It's
now the actual container name and the service name is available as
.ServiceName. For convenience .ServiceName is set to the container name
when using plain docker.

When support for Swarm is added the naming becomes inconsistent as the
name for the same concept changes between Swarm and non-Swarm. A compose
project is the equivalent of a Swarm namespace/stack. So the
.ComposeProject is changed to .Namespace as it's a quite neutral name
for both setups. The .ComposeColor is changed accordingly.

Similarly, in compose the container instance is called container number
whereas in Swarm it's called a slot. The new variable to be used for
both is .ContainerNumber.